Refactoriza flujo de creación de release en clase helpers
